requirements.txt
pil
south
django-markupfield
django-autoslug
django-sitetree
django-taggit
